A simple PHP script that measures the Collection
This calls the artworks endpoint (a page at a time) and collates the
longest edge to find out how long the Collection would be if you laid
everything down end to end, side to side, etc.

You need to replace the USERNAME and PASSWORD.

This is super hacky, my apologies.